move gql health check into module
